Poprawki podczas nagrania wideo
