+++ src/pl/dariusz202/controller/GameView.java
@@ -1,0 +1,7 @@
+package pl.dariusz202.controller;
+
+public enum GameView {
+
+    GAME, PLAY
+
+}
