+++ src/pl/dariusz202/model/Trainer.java
@@ -1,15 +1,20 @@
 // made by Wąsik Dariusz(github.com/dariusz202)
-package model;
+package pl.dariusz202.model;
 
 public class Trainer {
 
+    public static final int PLAYER_1_INDEX = 0;
+    public static final int PLAYER_2_INDEX = 1;
+
+    private int index;
     public String trainername;
     public int numberofpotion;
     public Pokemon[] pokemons;
     public int currentlypokemon;
     public int possiblerevive;
 
-    public Trainer(String trainername, Pokemon[] pokemons) {
+    public Trainer(int index, String trainername, Pokemon[] pokemons) {
+        this.index = index;
         this.trainername = trainername;
         this.numberofpotion = 3;
         this.pokemons = pokemons;
@@ -37,15 +42,19 @@ public class Trainer {
         System.out.println(this.trainername +" change Pokemon to "+pokemons[currentlypokemon].name);
     }
 
-    public void attack_other_trainer(Trainer other_trainer) {
+    public void attackOtherTrainer(Trainer otherTrainer) {
         Pokemon my_pokemon = this.pokemons[currentlypokemon];
-        Pokemon their_pokemon = other_trainer.pokemons[other_trainer.currentlypokemon];
+        Pokemon their_pokemon = otherTrainer.pokemons[otherTrainer.currentlypokemon];
         if(this.pokemons[currentlypokemon].current_health > 0) {
             System.out.println( "          " + this.trainername + " attack:");
             System.out.println( my_pokemon.name +" attack " + their_pokemon.name);
             my_pokemon.attack(their_pokemon);
         }
         else
-            System.out.println(this.trainername +" wanted to attack "+ other_trainer.trainername + " but his current pokemon is knocked out");
+            System.out.println(this.trainername +" wanted to attack "+ otherTrainer.trainername + " but his current pokemon is knocked out");
+    }
+
+    public int getIndex() {
+        return index;
     }
 }
